move code responsible for fetching mandatees to seperate file
